json output config
 - move ould classes to separate folder
