+++ config/output_wiring.json
@@ -1,0 +1,4 @@
+{
+  "piano": "Steinberg UR22mkII ",
+  "neuron_notes": "IAC Driver Bus 1"
+}
