+++ outputHandler.py
@@ -1,4 +1,4 @@
-import pygame.font
+# import pygame.font
 from numpy import intersect1d
 # import pygame.midi as pm
 
@@ -20,19 +20,13 @@ class OutputHandler(object):
         self.__output = outputs
         self.__input = {}
 
-        if Visuals.NAME in self.__output:
-            self.__output[Visuals.NAME].note_on(1)
-
         self.__now = time.time()
         self.__activeNotes = set()
         self.__neuron2NoteConversion = neuron2NoteConversion
 
-    def __setupInputs(self, inputList):
-        for name in inputList:
-            self.__input[name] = \
-                self.__getDevice(self.__name2Identifier[name], type='input')
-
     def update(self, fired):
+        if len(fired):
+            print("OutputHander fired", fired)
         neuron_ids = intersect1d(fired, self.pars['note_ids'])
         self.__checkKeyChange(neuron_ids)
 
@@ -50,16 +44,16 @@ class OutputHandler(object):
                  'tau_e', 'tau_i', 'cam_external_max'])
             pygame.display.update()
 
-    def turnOff(self):
-        for outputName in self.__output.keys():
-            if outputName == NeuronNotes.NAME:
-                self.__output[outputName].turnAllOff()
+    def turn_off(self):
+        for name in self.__output.keys():
+            if name == "neuron_notes":
+                self.__output[name].turn_all_off()
 
     def __checkKeyChange(self, neuron_ids):
         if len(neuron_ids) > 20:
             print(self.__output)
             self.__neuron2NoteConversion = (1 if self.__neuron2NoteConversion == 7 else 7)
-            self.__output[NeuronNotes.NAME].setNeuron2NoteConversion(
+            self.__output["neuron_notes"].setNeuron2NoteConversion(
                 self.__neuron2NoteConversion
             )
             # [output.setNeuron2NoteConversion(self.__neuron2NoteConversion) for
